# Improve performance of GCXS dot ndarray