@jcapriot 2024-02-17
> Thank you for the changes! Would you be willing to write benchmarks (examples in the `benchmarks` directory, see https://asv.readthedocs.io/en/stable/writing_benchmarks.html)

Sure!